Refactoring to untangle some of the modules (robots has now been split from stations module)
