+++ src/brom_drake/all/__init__.py
@@ -1,3 +1,4 @@
+from brom_drake.control.grippers.gripper_target import GripperTarget
 from brom_drake.DiagramTarget import DiagramTarget
 from brom_drake.DiagramWatcher import DiagramWatcher, DiagramWatcherOptions
 from brom_drake.file_manipulation.urdf import (
@@ -26,6 +27,7 @@ from brom_drake.PortWatcher.port_figure_arrangement import PortFigureArrangement
 from brom_drake.productions.ids import ProductionID
 from brom_drake.example_helpers import BlockHandlerSystem
 from brom_drake.file_manipulation.urdf import DrakeReadyURDFConverter, drakeify_my_urdf
+from brom_drake.robots.gripper_type import GripperType
 from brom_drake.utils.watcher import add_watcher, add_watcher_and_build, parse_list_of_simplified_targets
 from brom_drake.utils.leaf_systems import (
     BoolToVectorSystem,
@@ -46,6 +48,8 @@ __all__ = [
     'DrakeReadyURDFConverterConfig',
     'EndEffectorWrenchCalculator',
     'FigureNamingConvention',
+    'GripperTarget',
+    'GripperType',
     'MeshReplacementStrategy',
     'MeshReplacementStrategies',
     'parse_list_of_simplified_targets',
