+++ src/brom_drake/productions/pick_and_place/chem_lab3.py
@@ -32,7 +32,7 @@ from brom_drake.file_manipulation.urdf.simple_writer.urdf_definition import Simp
 from brom_drake.motion_planning.systems.open_loop_dispensers.open_loop_plan_dispenser import OpenLoopPlanDispenser
 import brom_drake.robots as robots
 from brom_drake.robots.gripper_type import GripperType
-from brom_drake.robots.stations.classical import UR10eStation
+from brom_drake.stations.classical import UR10eStation
 from brom_drake.productions import ProductionID
 from brom_drake.productions.roles import Role
 from brom_drake.productions.types import MotionPlanningAndGraspingProduction
