+++ src/brom_drake/robots/__init__.py
@@ -1,9 +1,7 @@
 from .gripper_type import GripperType
-from brom_drake.robots.stations.classical.ur10e import UR10eStation
 from .utils import find_all_link_names, find_base_link_name_in
 
 __all__ = [
     "find_all_link_names", "find_base_link_name_in",
     "GripperType",
-    "UR10eStation",
 ]
