--- src/brom_drake/robots/stations/classical/__init__.py
@@ -1,5 +1,0 @@
-from .ur10e import UR10eStation
-
-__all__ = [
-    "UR10eStation",
-]
