+++ src/brom_drake/stations/classical/__init__.py
@@ -1,0 +1,5 @@
+from brom_drake.stations.classical.ur10e import UR10eStation
+
+__all__ = [
+    "UR10eStation",
+]
