+++ tests/productions/roles/role_test.py
@@ -4,7 +4,7 @@ from pydrake.multibody.plant import MultibodyPlant, AddMultibodyPlantSceneGraph
 from pydrake.systems.framework import DiagramBuilder
 
 from brom_drake.example_helpers import BlockHandlerSystem
-from brom_drake.robots.stations.kinematic import UR10eStation
+from brom_drake.stations.kinematic import UR10eStation
 from brom_drake.productions.roles import Role, RolePortAssignment
 from brom_drake.productions.roles.role_port_assignment import PairingType
