+++ tests/robots/stations/classical/test_ur10e.py
@@ -18,7 +18,7 @@ from brom_drake.all import (
     add_watcher_and_build,
 )
 from brom_drake.robots.gripper_type import GripperType
-from brom_drake.robots.stations.classical.ur10e import UR10eStation
+from brom_drake.stations.classical.ur10e import UR10eStation
 from brom_drake.motion_planning.systems import (
     OpenLoopPlanDispenser
 )
